Highly-commented template
Wrote a bunch of quality of life code and created a project structure that's very easy to expand and add on to. A couple instructions:

If you want to create a new view (page), create a new Python file named after it within the "app" directory.  After following the same pattern laid out in "index.py," register the blueprint you just created within __init__ to allow for the URL association. Then create an HTML file that displays whatever you want for that particular URL, and throw it into "templates." Look at index.py - it should be pretty easy to follow.

The "static" directory is used for CSS files that define the overall look and feel of your project. You don't need to worry about it right now, but if you'd like some common patterns to appear in the front end - like the same font for divs - you can insert a style.css file and write down some rules. Again, don't worry about this for now.

You also want to add the instance folder to the gitignore after this. It contains an empty database, so pull this and ignore it from the repo after. You don't want ANYONE looking into your instance folder - once you develop to production, it'll contain very valuable information.

Hope this helps.